ref: Move utils in util/
